Backend/api/requirements.txt
blinker==1.6.2
click==8.1.7
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg==3.1.10
psycopg-binary==3.1.10
psycopg-pool==3.1.7
python-dotenv==1.0.0
typing_extensions==4.7.1
Werkzeug==2.3.7
